+++ CHANGELOG.rst
@@ -4,14 +4,15 @@ Change Log
 --------
 - remove the "set" / "change" aciton in the baselines with `utils/DeepQAgent`
 - stack multiple states in `utils/DeepQAgent`
-- leap net as a baseline
 - TrainingParam as json
+- save / load also this action space vector in default DeepQAgent.
 
 [0.3.0] - 2020-05-13
 --------------------
 - [ADDED] DeepQSimple, SAC and DuelQSimple baselines
 - [ADDED] utilitary code to create more easily agents
 - [ADDED] Multi processing training of agents using `grid2op.Environment.MultiEnvironment`
+- [ADDED] leap net as a baseline
 - [UDPATED] grid2op version `0.8.2`
 
 [0.1.1] - 2020-04-23